Massive refactor. In an effort to make symbooglix more like a library I have an IExecutorHandler interface which clients will be able to use for callbacks (once implemented).
This interface is used itself by the Executor. We've used a double
dispatch combined with a generic extension method that is attached
to all Absy classes. It would be nicer to attach it Cmd objects but
the Boogie class hierarchy is silly and makes it so that TransferCmd
does not inherit from Cmd. The implementation is concise but it may
bring potential performance problems!