Partially fix the horrific design of SMTLIBQueryPrinter. I've decided I don't care if the IExprVisitor methods are made public which simplifies things greatly. I've inverted things so now SMTLIBQueryPrinter inherits from IExprVisitor and has private nested class SMTLIBTraverser (inherits from Traverser) where as it was sort of the other way round before.
To allow a sensible assertion to continue to be used it was also
necessary to change the SimpleSMTLIBSolver so that the solver process is
created early so its TextWriter stream can be given to the
SMTLIBQueryPrinter. This does mean we are starting the solver process
earlier than its needed and possibly leaving around when it is never
needed again but hopefully this isn't much of an issue.