Fix terrible bug with requires statements.
Previously we treating requires statements on implementations like
assumes. This makes sense for the entry point because we want to
constrain the initial state. However for calls to other implementations,
requires statements need to be treated like asserts so we can follow
the path where the caller can't satisfy the requires statement and
produce an ExecutionState terminating in an error.

Now we treat requires differently depending if we are creating the
initial state or if we are performing a call.