Start refactoring how termination is handled. Termination is really a property of the ExecutionState so now there is a IterminationType property in ExecutionState that records this.
The termination type is deliberately an interface so different reasons
for termination can be added in the future.

This work required to teach ProgramLocation about other locations
in a Boogie Program.

The next task is to switch to an events based design for termination
(much like what is now used for break points).