Fix bug in Executor.EnterAnLeaveProcedure() where we try to continue executing (handle more requires or ensures) even if the current state had been terminated.
I've added a test case for this and also made
Executor.HandleAssertLikeCommand and Executor.HandleAssumeLikeCommand
make the CurrentState null when they terminate the current state so its
much more likely that we will crash if we do something wrong. The
motivation is because this bug was only caught by an assertion later on
during execution and I'd prefer if we caught it earlier.