Introduce TerminationTypeWithSatAndUnsatExpr and TerminationTypeWithUnsatExpr abstract classes which many TerminationTypes now descend from.
The motivation for this is that the Expr for a terminated state leading
to satisfiablity and/or unsatisfiability of constraints needs to be stored
so that is be used later on. The plan is to record this when logging a
terminated state's constraints to file. Currently the logging is
incomplete because the constraint that lead to termination is missing!