Fix coverage of assumeCmds following a GotoCmd when using the "Look ahead" goto implementation.
Before we would increment the coverage count of an assume even if we knew
following that path would be unsatisfiable. This would mean we would get
identical coverage reported using the naive or the "look ahead" goto
implementations.

However when using the "look ahead" goto implementation the counted
jumps would then appear inconsistent. For example we would report
that a certain Goto target was followed 4 times but the assumeCmd would
report it was visited 5 times!

This fixes that which makes Callgrind output look more sensible. However
this does mean that using different goto implementations will result in
different coverage being reported!