Fix DFSStateScheduler so it actually explores in DFS order. An interesting issue has arisen with how GotoCmds with > 2 targets are handled. With the current implementation of the DFSStateScheduler and Executor the State following the left most branch is visited first, then the other targets are visited right to left (there may be states in between) rather than left to right.
If we considered the ExecutionTree and we were to traverse it would
be like at each node visiting the left most child but then doing the
rest of the children right to left.

I'm not sure this is a problem. The Executor could be modified to insert
States into the scheduler in a different order but that seems wrong. The
Executor shouldn't need to care at all about how the IStateScheduler its
using is implemented.