Implement NonTerminatedStateRemoved event in the Executor. This will be used in the future to log information about non terminated states. A unit test was added to test this feature.
This unit test relies on the Executor timeout feature so a unit test
was added for that. Unfortunately the test didn't work correctly because
something was going wrong with calling Dispose() on the solver in
Executor.Terminate() so this has been removed.

We should tidy up the SimpleSMTLIBSolver at some point to remove
the code that supports being interrupted.