Change IStateScheduler's RemoveAll(Predicate<ExecutionState> p) to void RemoveAll()
The rationale for doing this that the previous interface pretty is sort
of tied to List<>. Although it's possible to implement using other data
structures its unncessary work because the method is currently not used.

We could of removed the method entirely but it might be useful as some
point to empty a state scheduler and emptying a container is usually
pretty trivial.