Change PassManager API. Now the Program is not taken in the constructor because there is no reason to tie a PassManager instance to a particular Program instance.
Instead the Run() method takes the Program. This allows a PassManager to
be reused on multiple programs. This isn't actually the motivation for
the design change. Instead this is needed to clean up the Executor API
so that a user cannot call Executor.Prepare() but is still able to give
the Executor a PassManager for use during preparation.