Try to teach the Executor to handle OutOfMemoryException more gracefully by catching the exception and if it's thrown then send the ExecutorTerminatedEvent (to allow loggers to run) and then rethrow the exception. This envolves a nasty hack which tries to make sure that the exception does not get thrown again by the listeners in the event. This works by preallocating a buffer of memory before hand and then freeing it when we run out of memory.
The most important result of doing this is that the
TerminationCounterLogger now records what states were found even if we
run out of memory.

During testing the CallGrindFileLogger, ExecutionTreeLogger and MemoryUsageLogger
had to be changed to be more robust because in really low memory
situations Executor.Prepare() might not finish and so the loggers would
crash because they expected preparation to be done when they perform
logging.