Add Executor.CheckUniqueVariableDecls property that disables the checking of the unique variable property. Generally this shouldn't be used but during my experiments it causes an issue with one of the benchmarks. If the constraint independence solver is used then the issue goes away but it wasn't being used for the snapshot I was working on.
The issue in particular was a boogie program that had a quantified axiom. Without
the constraint independence solver the quantified constraint will be used
in every query to the solver making every query fail (i.e. checking the
unique constraint would fail).

In general disabling the checking of the unique property (and the axioms
and the requires clauses) is only safe if it is known a priori that the
boogie programs given to Symbooglix has consistent assumptions.