Fix bug #8.
In GlobalDDE we would fail to remove dead axioms and functions if the
axiom also contained uses of interpreted functions (i.e. bvbuiltin and
builtins) and those interpreted functions were used in the program.

The original reason for this is behaviour is that we don't want to
remove axioms that constrain the behaviour of a used uninterpreted
function. However the built-ins are interpreted hence axioms won't
be changing how they are interpreted (i.e. constraining their
behaviour).

Now GlobalDDE no longer considers uses of interpreted functions
in axioms as "uses" which allows it remove more dead axioms and functions.