Change the IExprBuilder interface to take Triggers to the ``ForAll()`` and ``Exists()`` methods.
This required changing several related classes and tests. A few test
cases were added to check this behaviour. The DuplicationVerifier also
had to be taught to visit trigger expressions to properly check
duplication.

The motivation behind this change (apart from it being a long standing
FIXME) is so we can support passing user defined triggers to the
underlying solver.