Fix indexing into multi-arity maps.
This issue was reported by @sekunze in #18.

The problem was that the indicies that were being passed to MapProxies
were in the wrong order when multi-arity maps were being used which
caused a type checker exception to be raised when the multi-arity map
used different types.

A small test case has been included.